+++ src/features/basic/custom-item-sorting/tile-state.ts
@@ -2,6 +2,7 @@ import { getTileState as getBaseTileState } from '@src/store/user-data-tiles';
 
 export interface TileState extends UserData.TileState {
   activeSort?: string;
+  reverseSort: boolean;
 }
 
 export function getTileState(tile: PrunTile) {